+++ packages/client/src/types.ts
@@ -79,8 +79,8 @@ export type ViewFnApi<A extends Record<string, Record<string, any>>> = {
 export type TxApi<A extends Record<string, Record<string, any>>, Ext, Asset> = {
   [K in keyof A]: {
     [KK in keyof A[K]]: A[K][KK] extends {} | undefined
-      ? TxEntry<A[K][KK], K & string, KK & string, Ext, Asset>
-      : TxEntry<any, K & string, KK & string, Ext, Asset>
+      ? TxEntry<A[K][KK], Ext, Asset>
+      : TxEntry<any, Ext, Asset>
   }
 }
 
@@ -90,7 +90,7 @@ export type OfflineTxApi<
 > = {
   [K in keyof A]: {
     [KK in keyof A[K]]: A[K][KK] extends {} | undefined
-      ? OfflineTxEntry<A[K][KK], K & string, KK & string, Asset>
+      ? OfflineTxEntry<A[K][KK], Asset>
       : unknown
   }
 }