+++ packages/client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Removed `Transaction<Pallet, Name, Args, Asset, Ext>` pallet name and args generics to `Transaction<Asset, Ext>`. `decodedCall` is now a loose type.
+
 ### Removed
 
 - exported `TxCall`, `TxObservable`, `TxPromise` and `TxSign` interfaces, inlined directly into `Transaction`